去掉动画类名前缀，使用Float类型

+++ Source/AnimationKit/SpringAnimation.swift
@@ -10,11 +10,11 @@ import Foundation
 public class SpringAnimation: Animation {
     
     private var velocity: Any?
-    private var speed: CGFloat?
-    private var bounciness: CGFloat?
-    private var tension: CGFloat?
-    private var friction: CGFloat?
-    private var mass: CGFloat?
+    private var speed: Float?
+    private var bounciness: Float?
+    private var tension: Float?
+    private var friction: Float?
+    private var mass: Float?
     
     // MARK: - Override
     override func createAnimation(type: String, view: UIView) -> MLAValueAnimation {
@@ -25,42 +25,42 @@ public class SpringAnimation: Animation {
     
     // MARK: - Public
     @discardableResult
-    public func springVelocity(_ velocity: CGFloat) -> Self {
+    public func springVelocity(_ velocity: Float) -> Self {
         self.velocity = velocity
         return self
     }
     
-    public func springVelocity(_ velocity: Array<CGFloat>) -> Self {
+    public func springVelocity(_ velocity: Array<Float>) -> Self {
         self.velocity = velocity
         return self
     }
     
     @discardableResult
-    public func springBounciness(_ bounce: CGFloat) -> Self {
+    public func springBounciness(_ bounce: Float) -> Self {
         self.bounciness = bounce
         return self
     }
     
     @discardableResult
-    public func springSpeed(_ speed: CGFloat) -> Self {
+    public func springSpeed(_ speed: Float) -> Self {
         self.speed = speed
         return self
     }
     
     @discardableResult
-    public func springTension(_ tension: CGFloat) -> Self {
+    public func springTension(_ tension: Float) -> Self {
         self.tension = tension
         return self
     }
     
     @discardableResult
-    public func springFriction(_ friction: CGFloat) -> Self {
+    public func springFriction(_ friction: Float) -> Self {
         self.friction = friction
         return self
     }
     
     @discardableResult
-    public func springMass(_ mass: CGFloat) -> Self {
+    public func springMass(_ mass: Float) -> Self {
         self.mass = mass
         return self
     }
@@ -68,30 +68,30 @@ public class SpringAnimation: Animation {
     // MARK: - Private
     private func configAnimation(_ anim: MLASpringAnimation) {
         if let b = self.bounciness {
-            anim.springBounciness = b
+            anim.springBounciness = CGFloat(b)
         }
         if let s = self.speed {
-            anim.springSpeed = s
+            anim.springSpeed = CGFloat(s)
         }
         if let t = self.tension {
-            anim.dynamicsTension = t
+            anim.dynamicsTension = CGFloat(t)
         }
         if let f = self.friction {
-            anim.dynamicsFriction = f
+            anim.dynamicsFriction = CGFloat(f)
         }
         if let m = self.mass {
-            anim.dynamicsMass = m
+            anim.dynamicsMass = CGFloat(m)
         }
         if let v = self.velocity {
             switch v {
             case let v1 as CGFloat:
                 anim.velocity = v1
-            case let v2 as Array<CGFloat>:
+            case let v2 as Array<Float>:
                 switch v2.count {
                 case 1:
                     anim.velocity = v2[0]
                 case 2:
-                    anim.velocity = CGPoint(x: v2[0], y: v2[1])
+                    anim.velocity = CGPoint(x: CGFloat(v2[0]), y: CGFloat(v2[1]))
                 default: break
                 }
             default:
